Add timeout to framework shutdown! macro
This change adds a 5 minute timeout to the shutdown macro, in order to
fail any tests that have taken too long during that process.

This timeout is accomplished by passing the shutdown handle to a thread,
while a second thread sleeps for the timeout duration.  Both send a
message over a channel at completion, where the first to complete
"wins". If the timeout wins, it fails the test.

In order to make this change, it requires the macro to have only a
single shutdown handle passed to it.  As the only current uses of this
macro is to shutdown a network struct, the restriction currently has no
effect. In future commits, it may be possible to return it to have
multiples, but it requires a far more complex solution.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>