Replace sleep/shared state with channel
This change fixes a test that would sleep a very short amount of time
while checking the shared state to see if the data has arrived.  It
replaces the lock-wrapped vector with a channel.  The test will block
until it receives the value, allowing the dispatch thread time as much
time as it needs to process the message.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>