Use Bionic builders for GHA CI
"ubuntu-latest" GHA runners are based on Ubuntu Focal, which includes the
option "trust-dns" in /etc/resolv.conf.

Docker containers inherit DNS settings from the host, including the contents
of /etc/resolv.conf.

Through a chain of dependencies, version 0.6.3 of resolv-conf is being
used which contains a bug in parsing resolv.conf files containing the option
"trust-dns". This was fixed in 0.7.0 but we're unable to update without
updating actix.

This commit can be reverted once actix is updated to a newer version which
pulls in resolv-conf >= 0.7.0.

https://github.com/tailhook/resolv-conf/issues/25

Signed-off-by: Ryan Beck-Buysse <rbuysse@bitwise.io>