Ensure mesh ctrl chan is registered before return
This change fixes an issue where the control channel is possibly not
registered with the underlying mio::Poll before the Control and Incoming
structs are returned to the caller, due to thread contention.  This
leads to a scenario where a connection is added to the mesh over the
control channel immediately after it is created not triggering an event
via the Poll.

A particular suspicious "TODO" in the mio-extra's channel implementation
(see mio-extras/@f8c5df7 src/channel.rs, L286-287), there is a likely
race condition surrounding items being put on a mio channel before it is
registered.  This implies that the best course of action is to ensure
that the channel is registered before it is operated on.

Taking inspiration from the mio tests (see mio/@642e56d
test/test_custom_evented.rs), the fix involves using a Barrier to have
the spawning thread wait, if necessary, for the registration (which
happens within the Reactor constructor) to complete, before continue the
function.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>