Revert transact version to 0.2 instead of 0.3
The API for SmartContractArchive::from_scar_file changed
between the version enforcing different requirements of
the the used scar files.

To stay compatible with 0.4.1 splinter, the version is
moved back to 0.2. in the scabbard cli. This is does however
result in a different version of the transact error being
returned by the sabre sdk builders that are used to created
the transactions. As such the errors must be wrapped into
a different type.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>