Add tests for disbanding while nodes are stopped
This change adds integration tests around the process of disbanding a
circuit while the nodes receiving payloads are down. This allows us to
validate that payloads are handled as expected once a node is stopped
and restarted.

Currently, the tests added are ignored because of a known bug that
occurs when a node is restarted. At restart, the node's endpoint
changes, which disables the nodes from communicating as the circuit
definition holds the endpoint the node was initially started with.

Signed-off-by: Shannyn Telander <telander@bitwise.io>