Fix manual implementation of `Option::map` lint
This change fixes a lint introduced in the 1.52.0 release of Rust
flagging items like

    match Some(0) {
        Some(x) => Some(x + 1),
        None => None,
    };

Use instead:

    Some(0).map(|x| x + 1);

In some cases this requires an `as &(dyn Error + 'static)` as the option
includes a `Send` constraint as well.

In some cases this can be replaced with a `as_deref` call.

See for
    https://rust-lang.github.io/rust-clippy/master/index.html#manual_map
for details

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>