Add public key to circuit members in AdminServiceStore
This commit adds an optional public key to both ProposedNode and
CircuitNode. This public key will be used during challenge
authorization.

This change requires that the Circuit member list is updated to
contain CircuitNodes instead of just the node id. This will allow
storing the public key's directly in the circuit.

This is a breaking change against the YAML format, as such if a
node is using the YamlAdminServiceStore it cannot support
including public keys in the member list and thus will not
be able to support challenge authorization.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>