Update Rest API and CLI to return the updated member list
V2 Rest API has been updated to return the circuit nodes
with the public key if challenge authorization experimental features
is enabled.

The CLI has been updated to print the public key for the node when
showing a circuit and proposal.

Due to the changes to the circuit member list, these changes
have to happen in the same commit.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>