Change to_string_pretty to to_string in websocket.rs
to_string_pretty makes the message larger then to_string.
This was causing uploading smart contract state changes in
scabbard to be too large for the maximum size of the websocket
frame and caused the gameroom example to break.

to_string was used before but was accidently replaced in a
previous commit.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>