Add v1 Trust Peer Authorization
This commit includes the message handlers for the new authorization
message, as well as separating out the authorization state machine into
several state machines to make room for new actions/states in the new
trust procedure. This will make it easier in the future to add a
new state machine for challenge authorization.

The v0 Trust is still supported along with v1. If a ConnectRequest
is the first message received v0 trust will be used. If
AuthProtocolRequest is the first message v1 Trust will be used.
This is required to stay backwards compatible with 0.4.

It is also now required to track both our local authorization state
and the state of the remote as separate state machines. This is
because authorization will be happening in parallel.

The new implementation is behind a new experimental features
`trust-authorization` added to libsplinter and splinterd.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>