Update RefMap to have a generic key
The RefMap is used to keep track of peer's to know when
the connection should be dropped. With the addition of
challenge authorization the peer id is going to be updated
to be an Enum instead of a String, as such the reference map
needs to be updated to take keys that are other types.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>