Replace string peer id with PeerAuthorizationToken
This enables routing over the correct identity based on how the
connection was authorized. The layer that knows what authorization
type is required is the admin service, as such much of the public
API of the PeerManager, PeerManagerConnector, PeerInterconnect, and
network dispatchers had to be updated to use PeerAuthorizationToken
instead of a string.

The PeerID used by network dispatcher before wrapped a string but now
wraps a PeerAuthorizationToken. To make it easier to know which
PeerAuthorizationToken should be used in circuit handlers, a method
has been added to the routing table return the correct token based
on the authorization type used by the circuit.

This currently raises issues for two handler types. NetworkEcho knows
nothing about circuits and therefore cannot get the authorization type
required to send the message, so it is currently restricted to Trust.

Admin services that are using challenge authorization will have an
admin service id in the form admin::public_key::<public key as hex>
and it will be parsed by the AdminDirectMessageHandler. This will
work for the short term but before challenge authorization is stabilized
a better solution should be implemented.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>