Update handling of proposed circuit
This change updates the `handle_proposed_circuit` function within the
shared admin service. This update removes a section of the function that
added the proposed disband members to a list that was not used beyond
being added to. Rather, the function now checks each node ID in the list
of proposed disband members to verify that these nodes have an
agreed upon protocol version. If not, these disband members are added to
a list of proposed members that must go through protocol agreement.

Signed-off-by: Shannyn Telander <telander@bitwise.io>