Add PeerAuthorizationToken to peer module
When challenge authorization is implemented the Peer will need
to be able to be identified by a public key. Currently, the
peer is only referenced by a peer_id which corresponds to a
node_id, the identifier used for Trust. This token will be used
to replace the ID used for the PeerMangaer and PeerInterconnect.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>