Update main to use log-config if applicable
This commit adds the logic to conditionally read logging configs from
config toml file. There are three "happy" paths, a valid config with log
options, a valid config without any logging options, and log-config
being turned off.

If the config struct has logging options we deserialize and use those
configurations.

If log-config is enabled and the config struct has no log options we use
the defaults. This makes it backwards compatible with old config files.

If log-config is turned off we provide a reasonable default equivalent
to the current functionality.

If the config file can't be read, can't be deserialized, or is somehow
otherwise invalid we print the error and exit.

Signed-off-by: Caleb Hill <hill@bitwise.io>