Update abandon circuit to not be bound by protocol version
Before a circuit that was version 1 or if the agreed admin
protocol version was 1, a circuit could not be abandoned
because this operations is not supported in previous versions.

These checks have been removed. This will allow a 0.6 splinter
node to abandon a circuit that was created with a 0.4 node.
This is allowed because the operation does not require
confirmation from the 0.4 node. The 0.4 node will print an error
and ignore the notification sent when the circuit is abandoned.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>