Add list_nodes and get_node_token to PeerAuthorizationTokenReader
Also updates ListPeerAuthorizationTokens to be named
PeerAuthorizationTokenReader to be more accurate with the
new methods.

PeerAuthorizationTokenReader trait is a helper trait to get a list
of tokens from a circuit or proposal. There were several cases
that were not covered by the existing list_tokens method.

list_nodes returns the tokens along with other important pieces of
information about the node, such as its endpoints and its admin
service id.

get_node_token allows for getting a specific node's token from a
circuit. This is used to simplify getting the local nodes expected
authorization type.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>