Enable Challenge Authorization
This commit sets up the new challenge authorization message
handlers in create_authorization_dispatcher. To properly set
up the challenge authorization handlers information about the
expected authorization from the remote node, as well as what
authorization type the local node should use must be passed
at peer/connection request time. This information will be
known by the node adding the peer from the PeerManager
and must be passed down to the ConnectionManager and the
AuthorizationManager.

This information must also be stored to be used again on
connection restarts and trying new peer endpoints.

The admin service needs to be updated to be get the required local
authorization as well as updates the token to peer map to be
able to properly handle connection/disconnection notifications
as well as making it easier to map the token to the members
admin service.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>