Update inbound/outbound connection comparison to use correct identity
In the peer manager, it is possible that both sides could be trying
to connect to each other at the same time and result in duplicate
connections. The peer managers need to compare the two connections
and decide which one should be closed. Both peer managers need
to pick the same connection otherwise they can get into a loop
of peer reconnection. Note this only effects full peers, not
unreferenced.

With the switch to the PeerAuthorizationToken as the ID, the previous
solution of checking a local node id against the identity of the
connection no longer works if challenge authorization was used.

To fix this, the comparison is now done between the identity of the
new connected peer and the required local authorization type.

This commit also updates adds a default required local authorization
type for when challenge authorization is not enabled so the check
can be done in the same way.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>