Return AuthorizationError if msg is received out of order
The old authorization pattern included logging a warning if a
message was received out of order and ignored the message.
This left the connection being authorized in a odd state because
the connection would never be authorized or unauthorized.

This commit changes that process to return an AuthorizationError and
unauthorizes the connection.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>