Add deserializable log config structs
To make the config file look how we want it to there needs to be an intermediate
structure between the file and our config structures. The general file structure
we are aiming for is

[appenders/loggers]

name = {<structure fields>}

Into implementations between (name, struct_without_name) to full_struct_with_name
are the glue between publicly accessible configuration options and our internal
structures.

Signed-off-by: Caleb Hill <hill@bitwise.io>