Make main use logging configurations
Lots of things can go wrong before we fully read in all configuration values
including the logging configurations. Those errors need to be visible even if
the logging configs haven't been applied yet or are ultimately unable to be
applied.

The solution here is just a two part log initialization. We provide reasonable
defaults that are always applied. Then once the Config object has been created
we update the running log config with the users preferences.

Signed-off-by: Caleb Hill <hill@bitwise.io>