Add node_id upgrade file to database functionality
UpgradeAction is for all upgrade tasks and moving the systems node_id into the
database qualifies as an upgrade.

The node_id is only written to the database if:

- The node-id-upgrade feature is enabled
- The database connection string works
- The node_id file is not empty
- There isn't already a node_id in the db

Otherwise it throws an error because the prereqs are not met. If all those
things are true it will rename the node_id file to node_id.old .

Signed-off-by: Caleb Hill <hill@bitwise.io>