Enforce at least one signer in authorization handler
Before if no signers were provided, challenge authorization
would be disabled. This commit changes the builder for the
AuthorizationDispatchBuilder to raise an en error if a
at least one signer is not provided.

This commit removes challenge_configured from
AuthProtocolRequestHandler. challenge_configured was set
to false if challenge-authorization was enabled but no
signers were configured. This is no longer a valid state.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>