Add `/biome/keys` test and add `PostKey` derives
Add a new unit test that validates that the PUT `/biome/keys` endpoint
functions as expected.

Update `PostKey` so it derives from `Debug`, `PartialEq`, and `Eq`. This
is so it can be easily compared against other `PostKey` structs, and the
results of that comparison can be displayed in unit tests.

Signed-off-by: Lee Bradley <bradley@bitwise.io>