Replace MetricsError with InternalError
Originally MetricsError was a wrapper around an InternalError but was
only used within metrics related code. InternalError is more general in
that it can logically be used anywhere MetricsError was but also in code
not related to metrics. This simply removes the MetricsError variant and
replaces it with the InternalError variant.

Signed-off-by: Caleb Hill <hill@bitwise.io>