Update --node-public-key to work with --node-file
Previously --node-public-key values were only being added
to the nodes defined by --node, however if using challenge
authorization public keys need to be added for nodes that
were added using a --node-file as well.

The node file format is not going to be changed to support adding
the challenge authorization keys directly, so they must be provided
separately on the command line.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>