Add splinterd biome tests and update `NewKey`
Add splinterd tests to validate the biome client functions correctly
when integrated. Add `biome-replace-keys` feature to splinterd to
trigger testing behavior.

Update `NewKey` so it derives from `Clone`, `PartialEq`, and `Eq`. This
is so it can be cloned and easily compared against other `NewKey`
structs. Currently this is used in the above mentioned tests.

Signed-off-by: Lee Bradley <bradley@bitwise.io>