Fix warnings that were being emitted from biome test
The warnings were being raised because several fields in
the test specific biome items were not being read. But
by keeping the fields in the structs the fields can still
be checked that they are in the returned response.

This commit puts an underscore in front of the unused fields
to remove the warning and adds a serde rename so the response
can still be deserialized correctly.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>