Change error handling to be less extreme to user
Printing a message starting with "ERROR" for somewhat expected outcomes
isn't user friendly. There are a few reasons node_id and yaml actions
can fail which may not really be traditional errors. For example the node_id
action will fail if it has been run before but the upgrade action can
and should continue without issue.

This change still prints the underlying error but includes that the
action is simply skipped and thus may not be a real error/problem.

Signed-off-by: Caleb Hill <hill@bitwise.io>