Update test broken by changes to return types
Making the error handling of the upgrade command simpler also changes
some branches return types from Err to Ok. The idea behind those changes
is that the files not being there isn't necessarily an error and might
just indicate the upgrade command has been run before.

So to that effect it logs a warning message and returns Ok. The test has
been updated to check it returns an Ok(()) instread of
an Err(_).

Signed-off-by: Caleb Hill <hill@bitwise.io>