Improve `EnvPartialConfigBuilder` testability
EnvPartialConfigBuilder could not previously be tested, as the tests
depended on setting and fetching environment variables in a way that
would inadvertantly affect other tests.

This was also problematic because higher level integration tests that
depended on EnvPartialConfigBuilder inherited the testability issues.

This change modifies EnvPartialConfigBuilder to abstract any
dependencies that could affect other tests. Environment variables are
now retrieved through a generic EnvStore that can be swapped out with
a mocked enviroment variable store.

Signed-off-by: Lee Bradley <bradley@bitwise.io>