Update connection manager to allow multiple connections for endpoint
Before challenge authorization, there was only 1 connection allowed
for an endpoint.

With challenge authorization, it is possible to have more then one
connection connected to the same endpoint if the peer authorization
token is different.

This commit updates the connection manager to use the connection id
as the unique identifier instead of the endpoint. This also requires
adding the connection id to remove_connection and updating the
peer manager to pass the connection id on remove.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>