Update PeerManager to support multiple peers having the same endpoint
Before challenge authorization was implemented, endpoints to peer id
were expected to be 1:1. Now that is no longer true because the same
peer may support multiple peer ids with different authorization
types that have the same endpoints.

This commit updates the peer manager and the peer map to remove enforcement
of no duplicate endpoints.

Also adds a check to make sure that two different trust peer ids do not have
the same endpoints, this should not be allowed.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>