Add test for creating 2 circuits with completly different keys
The existing test creating 2 circuit where one node has the same
key in each circuit and the other has a different key in each.
This test still does not pass and is still ignored.

This commit adds a second test that sets both nodes keys to
different keys in each circuit. This test does pass.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>