Add local authorization type to InboundConnection notification
Because peer connections need to be identified by both the peer's
authorization as well as how the local node was authorized, inbound
connections need to return this information in the notfication.

This change updates authorization, as well as connection manager to
return the local authorization from inbound connections.

The InprocAuthorizer constructor has been updated to require the
node ID so it can be added as the local authorization.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>