Add receipt_store to ScabbardState
Add a `receipt_store` field to the `ScabbardState` struct that has type
`Arc<RwLock<dyn ReceiptStore>>` and is behind the experimental feature
"diesel-receipt-store". This `ReceiptStore` is the new receipt store
trait implemented with postgreSQL and SQLite backends in Sawtooth-lib.

The receipt store is created using the receipt_db_url passed into
ScabbardState's `new` method.

Add a #[cfg(not(feature = "diesel-receipt-store"))] feature guard to
the `transaction_receipt_store` field so that the lmdb receipt store
database is not created if the diesel-receip-store feature is
enabled.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>