Add a second `Events` implementation
Put the current `Events` struct and implementation behind a not
"diesel-receipt-store" feature guard so that it is not used when
"diesel-receipt-store" is enabled.

Create a second events struct and implementation that uses the diesel
backed receipt store behind the experimental feature
"diesel-receipt-store".

Update the corresponding tests and methods that use `Events`.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>