Replace PeerAuthorizationToken with PeerTokenPair
This commit replaces the use of PeerAuthorizationToken for the peer
with a PeerTokenPair which contains both the peer's authorization
along with the local nodes authorization type. This is required
because a node should be able to connect to another node over
different connections using different public keys.

The PeerManager, PeerMap, PeerConnector and PeerInterconnect have
all been updated to use PeerTokenPair as the Peer ID for the peers.

Dispatcher PeerID has been updated to use PeerTokenPair along with
all of the message handlers.

Finally, the admin service has also been updated to user PeerTokenPair.
This required an update the admin service hack so the admin service id
has been updated to include both the peers public key and the local
nodes id "admin::public_key::PUBLIC_KEY::public_key::PUBLIC_KEY"

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>