Update metrics feature to use v0.17 of metrics in libsplinter
lipsplinter was using v0.12 of the metrics library. Before
stabilization we want to update it to use the newest version
which is currently 0.17.

This commit updates the API to match the changes made in the new
versions which include the following:
  - Change what is passed to trait methods
  - Adds handler for GaugeValue and keeps track of existing gauge values
    similar to counter
  - Replace use of Cow with metrics::SharedString (wrapper around Cow)
  - Implement register methods on the Recorder trait

Finally, the metrics features and module has been renamed to tap, like
tap a tree. This is required because previously the metrics crate
was renamed in libspinter to metrics_lib to differentiate between
the local metrics and the crate. However in 0.17 the macros have
been moved to their own crate and reference the metrics crate within
the macros which breaks the rename. Therefore, we must rename the
feature/module within libsplinter.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>