Properly send protocol request before vote/disband
If a service protocol is missing when receiving a vote or
a disband request, a service protocol request should be sent
to the peer. This was being done but was not accounting for
updating the message sender to match the admin service
id if challenge authorization was used.

This commit sets the sender to the local peer id, which
should be the reverse of the recipient.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>