Pull storage creation up to ScabbardFactory
This change pulls the creation of the receipt store up to the factory
level. It also configures the state storage paths at the same time.
This removes the requirement that the service knows anything about its
internal storage configuration.

The configuration is provided via the ScabbardStorageConfiguration enum,
which is validated at factory-build time.  It's converted into an
internal storage config enum that will be used when the factory creates
a service.

It also requires that the service is provided a closure that handles the
purge functionality.  This functionality changes based on the underlying
storage configuration - LMDB files are deleted, but Diesel-based
solutions don't provide any clean-up API's.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>