Add MEMORY constant to sqlite db access code
":memory:" is the connection string for memory backed sqlite instances.
The code checks in multiple places if the connection string is ":memory:".
This can be error prone and its kind of a magic value so its a good candidate
to become a const.
This commit makes a const and replaces all usages of ":memory:" within the
file.

Signed-off-by: Caleb Hill <hill@bitwise.io>