Add node_id access logic to daemon
node_id's are in the database now so accessing those values is more
complex than should be in main.
At initilization time the daemon might not know the node_id unless its
been passed in on the cli. The `get_node_id` function determines if the
cli_node_id matches the database_node_id match and which one should be
used.

Signed-off-by: Caleb Hill <hill@bitwise.io>