Move logging::config to config::logging
Configuration is part of the splinterd's API to the user. Changes to the
configuration (such as changes to toml files) must be obvious and
intentional. This is only possible if the configuration code exists in
isolation from non-configuration code. Thus, the config module is
isolated and not allowed to import pieces of its implementation from
other splinterd modules.

The logging code previously violated this principle, and this change
is intended to place all of the logging configuration code within the
config module.

Signed-off-by: Shawn T. Amundson <amundson@bitwise.io>