Check protocol version against proposal when validating vote
It is possible that a proposal could have been stored in state
and when a vote has been received the agreed upon protocol
version has changed. This is possible due to node restarts/upgrades.

If the agreed upon protocol no longer supports the proposal,
the vote is rejected.

In a future commit, this proposal should also be removed from state.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>