Put off removing peers when a proposal is removed
If the peers are dropped immediately when the proposal is
removed, the messages that finish consensus on maybe dropped
before receipt.

This commit instead moves those peer refs into a temporary vec,
that includes an Instant for when they are added to the list.
After 10 seconds, the peers can be dropped. This is triggered by
either propose_change or create_proposal, both functions being
called by consensus. This ensures that the peers are not
dropped until the proposal has been handled.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>