Enable and fix reject integration tests
After saving off the peer refs to be removed at some
time in the future when an associated proposal is
rejected/removed the rejected integration tests can
now be passed. This is because message will not be
dropped allowing all nodes to finish consensus.

Turning on the integration tests showed a few small bugs
that were fixed to get all the tests passing.

The tests that include restarting are still blocked by
fixing the endpoint change.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>