Add Mocked NodeIDStores for testing
Testing with actual databases/files is more of an integration test than
a unit test. These NodeIDStores exist to make testing the import_store
function easier.
One functions as a regular store in memory that can be writen to and
read from. The other always returns an error. Together it makes it
possible and relatively straight forward to test all return combinations
for import_store.

Signed-off-by: Caleb Hill <hill@bitwise.io>