Add scabbard upgrade module
This change adds a function to upgrade all of the services such that the
commit hash is moved from the lmdb database to the configured diesel-
backed CommitHashStore.

This upgrade function requires its own set of Upgrade stores as they

1. Need to connect to the same database pool
2. Include a factory fn for constructing CommitHashStores, which are a
   scabbard concept.

As this process can produce both InvalidState- and InternalErrors during
its process, it includes an UpgradeError type. This error has been
placed in a new error module.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>