Handle empty registry yaml
This change fixes the local YAML registry such that it can handle an
empty file, as well as an empty YAML array. These are distinctly
different as one is zero bytes, the other is a fix set of YAML symbols.

Without this change, if a registry file is empty at the start of
splinterd (i.e. it was created before the first run), it will not see
any changes if nodes are added to the file while the node is running.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>