Update unreferenced peers to properly switch connections if duplicate
Previously, if there was an unreferenced peer connection when one
previously existed, it would always bee replaced.

This is a bad pattern if one side thinks the connection is a full
peer and might remove the incorrect side. This commit updated the
logic so unreferenced peer follow the same logic as full peers
when removing a duplicate connection.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>