Add MerkleState wrapper
This change adds a MerkleStateWrapper which allows the implementation to
be switched based on the internal implementation - either LMDB or
SqlMerkleState.

The primary reason this is handled as an enum, over a boxed trait is
that the current traits provided by transact are Read, Write, and Prune.
There is no single trait that includes all of these traits, so a wrapper
was necessary.

This wrapper may be removed with future version of transact where either
there is a single State trait, or the MerkleState and SqlMerkleState
implementations are merged.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>