Add From/TryFrom impls for toml->internal structs
To convert from how the configuration data is stored in the toml file to
how it is actually used in the code requires some "glue" trait
implementations to handle the transitions.

The basic idea is:

Toml file
    |
    v
Toml structs
    |
    v
Internal repesentation structs
    |
    v
A real full configuration.

This commit helps with the second arrow in the chain.

Signed-off-by: Caleb Hill <hill@bitwise.io>