Enable all stores to use database impls
This change modifies the Scabbard factory such that it can generate the
appropriate store types and use it to construct a scabbard service.

This includes a change in how the factory is configured.  The stores
that support transaction processing - `ReceiptStore` and
`CommitHashStore` - are solely configured to use a diesel-backed
instance.

The merkle state, however, may still be configured to use LMDB.  If this
is not set, then the merkle state is configured to use the database.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>