Fix type for admin service event id for postgres
For postgres, there is a difference between Integer and
BigInt. Before the event Id's were being set to BigInt
(using BigSerial) in the admin_service_event but Integer
on all other tables. However, these types need to be the
same so this commit adds a migration that changes the type
from Integer to BigInt.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>