Add enable_lmdb_state to Daemon
This change adds a flag to enable the use of LMDB for scabbard state,
when the "scabbard-database-support" feature is enabled.

When this flag is enabled, the scabbard factory is configured with an
LMDB state directory.  When the flag is disabled, scabbard stores
everything, including merkle state in the database specified by the DB
URL.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>