Add `get_batch_status`
Add an arg `get_batch_status` to `add_workload` and pass this arg's
value to the WorkerBuilder `get_batch_status` method. This arg is a
boolean value that determines if the worker will check the status of
batches after they are submitted to determine if the expected outcome
was acheived.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>