Update admin register route to handle protocol versions
When submitting the request the SplinterProtocolVersion will
determine the version of admin events that will be returned over
the event subscriber.

If the protocol is set to 1, 0.4 equivalent AdminEvents are
returned.

If the protocol is set to 2 or not set, 0.5/0.6 AdminEvent
messages will be returned.

This allows for backwards compatibility with 0.4 clients to
still interact with 0.5 splinter nodes.

Signed-off-by: Andrea Gunderson <agunde@bitwise.io>