Update `NodeBuilder` to have `permission_config`
Update the `NodeBuilder` to have a `permission_config` field with type
`Option<Vec<PermissionConfig>>` if any permission configurations are
stored in this field they are used to create roles and assignments in
the role based auth store. A `RoleBasedAuthorizationHandler` is created
with the role based authorization store after the roles and assignments
are set.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>