Add `commit_2_party_circuit_with_auth` method
Add a `commit_2_party_circuit_with_auth` method which takes an `auth`
arg and passes it to `admin_service_client_with_auth` and
`admin_service_event_client_with_auth` when retrieving the clients.
The clients use the auth string in the authorization header when
making requests to the REST API. This additional method allows for
circuits to be created when the network is configured with
authorization handlers other than the default mock authorization
handler.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>